requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Deprecated==1.2.13
Django==4.2.1
djangorestframework==3.14.0
gcloud==0.18.3
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
jwcrypto==1.4.2
Markdown==3.4.3
oauth2client==4.1.3
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.17
pyparsing==3.0.9
Pyrebase4==4.7.0
python-jwt==4.0.0
pytz==2023.3
requests==2.30.0
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.2
wrapt==1.15.0
